feat(detail page): bug fixes in model delete, in model dircetory, and in...

Closes #11

See merge request domanspr/dnn_nas_and_pruning!14

+++ code/NASO/runs/views/ListRuns.py
@@ -38,13 +38,14 @@ def delete_autokeras_run(request, pk):
     # Fetch the object or return a 404 response if it doesn't exist
     obj = get_object_or_404(AutoKerasRun, pk=pk)
 
+    # Delete the folder:
+    if len(obj.model.directory) > 0:
+        folder = "auto_model/" + obj.model.directory
+        if os.path.exists(folder):
+            shutil.rmtree(folder)
+
     # Delete the object
     obj.delete()
-
-    # Delete the folder:
-    folder = "auto_model/" + obj.model.directory
-    if os.path.exists(folder):
-        shutil.rmtree(folder)
 
     # Return a JSON response to indicate successful deletion
     return JsonResponse({"message": "Object deleted successfully", "id": pk})
